+++ docs/source/api_references/roles.rst
@@ -1,0 +1,6 @@
+============================
+Role Utilities API Reference
+============================
+
+.. automodule:: toolbox.roles
+   :members:
